DGS-19203 refactor repo to add tests to schema registry workspace
