+++ schemaregistry/Makefile
@@ -9,8 +9,8 @@ DOCKER ?= ./run_docker_schemaregistry.sh
 
 # Paths
 SRC_DIR = .
-SR_TEST_DIR = ../test/schemaregistry
-DEK_TEST_DIR = ../test/schemaregistry/rules/encryption/dekregistry
+SR_TEST_DIR = ./schemaregistry/test
+# DEK_TEST_DIR = ../test/schemaregistry/rules/encryption/dekregistry
 INTEG_DIR = ../e2e/schemaregistry
 
 # Tasks
@@ -22,7 +22,7 @@ lint:
 	$(ESLINT) $(SRC_DIR) $(TEST_DIR) $(INTEG_DIR) $(DEK_TEST_DIR)
 
 test:
-	$(JEST) $(SR_TEST_DIR) $(DEK_TEST_DIR)
+	$(JEST) $(SR_TEST_DIR)
 
 integtest:
 	$(DOCKER)
