+++ schemaregistry/test/dekregistry/mock-dekregistry-client.spec.ts
@@ -1,5 +1,5 @@
-import { Dek, Kek } from "../../../../../schemaregistry/rules/encryption/dekregistry/dekregistry-client";
-import { MockDekRegistryClient } from "../../../../../schemaregistry/rules/encryption/dekregistry/mock-dekregistry-client";
+import { Dek, Kek } from "../../rules/encryption/dekregistry/dekregistry-client";
+import { MockDekRegistryClient } from "../../rules/encryption/dekregistry/mock-dekregistry-client";
 import { beforeEach, afterEach, describe, expect, it, jest } from '@jest/globals';
 import { TEST_KEK, TEST_KEK_NAME, TEST_KMS_TYPE, TEST_KMS_KEY_ID,
   TEST_KMS_PROPS, TEST_DOC, TEST_DEK, TEST_DEK_V2, TEST_ALGORITHM,
