+++ schemaregistry/test/oauth-client.spec.ts
@@ -1,4 +1,4 @@
-import { OAuthClient } from '../../schemaregistry/oauth/oauth-client';
+import { OAuthClient } from '../oauth/oauth-client';
 import { ClientCredentials, AccessToken } from 'simple-oauth2';
 import { beforeEach, afterEach, describe, expect, it, jest } from '@jest/globals';
 import * as retryHelper from '@confluentinc/schemaregistry/retry-helper';
